requirements.txt
PyPDF2==3.0.1
langchain==0.0.348
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
numpy==1.26.2
huggingface_hub==0.19.4